# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Furniskuy APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/barang-masuks": {
            "get": {
                "tags": ["BarangMasuk"],
                "summary": "getBarangMasukList",
                "description": "Get all BarangMasuks",
                "operationId": "7cb18ad123c77c2cc40ec569c973e9c1",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BarangMasuk"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["BarangMasuk"],
                "summary": "createBarangMasuk",
                "description": "Create BarangMasuk",
                "operationId": "ed27426d2514a9147dbc08d682294a39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BarangMasuk"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BarangMasuk"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/barang-masuks/{id}": {
            "get": {
                "tags": ["BarangMasuk"],
                "summary": "getBarangMasukItem",
                "description": "Get BarangMasuk",
                "operationId": "5e8b121a1bda7e1f9e2433d4202ef8c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of BarangMasuk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BarangMasuk"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["BarangMasuk"],
                "summary": "updateBarangMasuk",
                "description": "Update BarangMasuk",
                "operationId": "9e9fb2a8ebeb2a11ba29ab22b09a05ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of BarangMasuk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BarangMasuk"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BarangMasuk"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["BarangMasuk"],
                "summary": "deleteBarangMasuk",
                "description": "Delete BarangMasuk",
                "operationId": "90db3b83612b18001f160803e15141d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of BarangMasuk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BarangMasuk": {
                "required": [],
                "properties": {
                    "nama": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "tanggal": {
                        "description": "",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}